CVSL-1443 Reduce spam to govuk bankholiday endpoint
Recent changes mean that we are bypassing our cache when requesting bankholiday. This is currently happening twice per licence load.
