+++ src/test/kotlin/uk/gov/justice/digital/hmpps/createandvaryalicenceapi/service/LicenceServiceTest.kt
@@ -83,7 +83,7 @@ class LicenceServiceTest {
   private val auditEventRepository = mock<AuditEventRepository>()
   private val notifyService = mock<NotifyService>()
   private val omuService = mock<OmuService>()
-  private val bankHolidayService = mock<BankHolidayService>()
+  private val releaseDateService = mock<ReleaseDateService>()
 
   private val service = Mockito.spy(
     LicenceService(
@@ -98,7 +98,7 @@ class LicenceServiceTest {
       auditEventRepository,
       notifyService,
       omuService,
-      bankHolidayService,
+      releaseDateService,
     ),
   )
